requirements.txt
streamlit
python-dotenv
pytube
google-generativeai
youtube-transcript-api
